[ISPN-7535] Cache creation requires specific permissions when using security manager
Move Util.getClassLoaders() and ReflectionUtil.invokeAccessibly() to doPrivileged() when security manager is enabled
