+++ commons/src/main/java/org/infinispan/commons/util/ReflectionUtil.java
@@ -2,7 +2,6 @@ package org.infinispan.commons.util;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -163,17 +162,7 @@ public class ReflectionUtil {
     * @param parameters parameters
     */
    public static Object invokeAccessibly(Object instance, Method method, Object[] parameters) {
-      try {
-         method.setAccessible(true);
-         return method.invoke(instance, parameters);
-      } catch (InvocationTargetException e) {
-         Throwable cause = e.getCause() != null ? e.getCause() : e;
-         throw new CacheException("Unable to invoke method " + method + " on object of type " + (instance == null ? "null" : instance.getClass().getSimpleName()) +
-                                        (parameters != null ? " with parameters " + Arrays.asList(parameters) : ""), cause);
-      } catch (Exception e) {
-         throw new CacheException("Unable to invoke method " + method + " on object of type " + (instance == null ? "null" : instance.getClass().getSimpleName()) +
-               (parameters != null ? " with parameters " + Arrays.asList(parameters) : ""), e);
-      }
+      return SecurityActions.invokeAccessibly(instance, method, parameters);
    }
 
    public static Method findGetterForField(Class<?> c, String fieldName) {
