+++ commons/src/main/java/org/infinispan/commons/util/Util.java
@@ -128,33 +128,11 @@ public final class Util {
 
    public static ClassLoader[] getClassLoaders(ClassLoader appClassLoader) {
       if (isOSGiContext()) {
-         return new ClassLoader[] { appClassLoader,   // User defined classes
-               getOSGiClassLoader(), // OSGi bundle context needs to be on top of TCCL, system CL, etc.
-               Util.class.getClassLoader(),           // Infinispan classes (not always on TCCL [modular env])
-               ClassLoader.getSystemClassLoader(),    // Used when load time instrumentation is in effect
-               Thread.currentThread().getContextClassLoader() //Used by jboss-as stuff
-         };
+         return SecurityActions.getOSGIContextClassLoaders(appClassLoader);
       } else {
-         return new ClassLoader[] { appClassLoader,   // User defined classes
-               Util.class.getClassLoader(),           // Infinispan classes (not always on TCCL [modular env])
-               ClassLoader.getSystemClassLoader(),    // Used when load time instrumentation is in effect
-               Thread.currentThread().getContextClassLoader() //Used by jboss-as stuff
-         };
+         return SecurityActions.getClassLoaders(appClassLoader);
       }
    }
-
-    private static ClassLoader getOSGiClassLoader() {
-        // Make loading class optional
-        try {
-            Class<?> osgiClassLoader = Class.forName("org.infinispan.commons.util.OsgiClassLoader");
-            return (ClassLoader) osgiClassLoader.getMethod("getInstance", null).invoke(null);
-        } catch (ClassNotFoundException e) {
-            // fall back option - it can't hurt if we scan ctx class loader 2 times.
-            return Thread.currentThread().getContextClassLoader();
-        } catch (Exception e) {
-            throw new RuntimeException("Unable to call getInstance on OsgiClassLoader", e);
-        }
-    }
 
    /**
     * <p>
